[rush-serve] Add dependencies to operations, support abort
