+++ common/changes/@microsoft/rush/rush-serve-dependencies_2025-09-16-23-56.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@microsoft/rush",
+      "comment": "[rush-serve-plugin] Support aborting execution via Web Socket. Include information about the dependencies of operations in messages to the client..",
+      "type": "none"
+    }
+  ],
+  "packageName": "@microsoft/rush"
+}
