+++ rush-plugins/rush-serve-plugin/src/RushProjectServeConfigFile.ts
@@ -8,35 +8,9 @@ import { Async } from '@rushstack/node-core-library';
 import type { ITerminal } from '@rushstack/terminal';
 import { RigConfig } from '@rushstack/rig-package';
 import type { RushConfigurationProject } from '@rushstack/rush-sdk';
+
 import rushProjectServeSchema from './schemas/rush-project-serve.schema.json';
-
-export interface IRushProjectServeJson {
-  routing: IRoutingRuleJson[];
-}
-
-export interface IBaseRoutingRuleJson {
-  servePath: string;
-  immutable?: boolean;
-}
-
-export interface IRoutingFolderRuleJson extends IBaseRoutingRuleJson {
-  projectRelativeFile: undefined;
-  projectRelativeFolder: string;
-}
-
-export interface IRoutingFileRuleJson extends IBaseRoutingRuleJson {
-  projectRelativeFile: string;
-  projectRelativeFolder: undefined;
-}
-
-export type IRoutingRuleJson = IRoutingFileRuleJson | IRoutingFolderRuleJson;
-
-export interface IRoutingRule {
-  type: 'file' | 'folder';
-  diskPath: string;
-  servePath: string;
-  immutable: boolean;
-}
+import type { IRushProjectServeJson, IRoutingRule } from './types';
 
 export class RushServeConfiguration {
   private readonly _loader: ProjectConfigurationFile<IRushProjectServeJson>;
